requirements.txt
gitdb==4.0.9
GitPython==3.1.24
smmap==5.0.0
tqdm==4.62.3
typing-extensions==3.10.0.2
